+++ src/main/java/wooteco/subway/path/domain/strategy/additional/AgeDiscountPolicy.java
@@ -2,7 +2,7 @@ package wooteco.subway.path.domain.strategy.additional;
 
 public abstract class AgeDiscountPolicy {
     public final int calculateFare(int fare) {
-        return (int)(fare * discountRate());
+        return (int) (fare * discountRate());
     }
 
     protected abstract double discountRate();
