+++ src/test/java/wooteco/subway/path/domain/FareTest.java
@@ -4,7 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import wooteco.subway.path.domain.strategy.DistanceAdditionPolicy;
+import wooteco.subway.path.domain.strategy.discount.DistanceAdditionFactory;
+import wooteco.subway.path.domain.strategy.discount.DistanceAdditionPolicy;
 import wooteco.subway.path.domain.strategy.additional.AgeDiscountFactory;
 import wooteco.subway.path.domain.strategy.additional.AgeDiscountPolicy;
 import wooteco.subway.path.domain.strategy.additional.NoDiscount;
